+++ source/standalone_examples/api/isaacsim.replicator.examples/sdg_getting_started_04.py
@@ -67,7 +67,7 @@ def run_example():
 
     # Write data using the basic writer with the rgb and bounding box annotators
     writer = rep.writers.get("BasicWriter")
-    out_dir = os.getcwd() + "/_out_basic_writer_sim"
+    out_dir = os.path.join(os.getcwd(), "_out_basic_writer_sim")
     print(f"Output directory: {out_dir}")
     writer.initialize(output_dir=out_dir, rgb=True, semantic_segmentation=True, colorize_semantic_segmentation=True)
     writer.attach(rp)