+++ source/extensions/isaacsim.replicator.behavior/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## [1.1.4] - 2025-05-12
+### Changed
+- Using os.path.join to create output directory in sdg tests
+
 ## [1.1.3] - 2025-05-10
 ### Changed
 - Enable FSD in test settings