+++ source/extensions/isaacsim.replicator.examples/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## [1.1.16] - 2025-05-12
+### Changed
+- Using os.path.join to create output directory in sdg tests
+
 ## [1.1.15] - 2025-05-10
 ### Changed
 - Enable FSD in test settings